+++ RxCocoa/iOS/DataSources/RxTableViewReactiveArrayDataSource.swift
@@ -46,7 +46,7 @@ class RxTableViewReactiveArrayDataSourceSequenceWrapper<S: Sequence>
     , RxTableViewDataSourceType {
     typealias Element = S
 
-    override init(cellFactory: CellFactory) {
+    override init(cellFactory: @escaping CellFactory) {
         super.init(cellFactory: cellFactory)
     }
 
@@ -80,7 +80,7 @@ class RxTableViewReactiveArrayDataSource<Element>
 
     let cellFactory: CellFactory
     
-    init(cellFactory: CellFactory) {
+    init(cellFactory: @escaping CellFactory) {
         self.cellFactory = cellFactory
     }