+++ RxCocoa/iOS/UITextView+Rx.swift
@@ -65,7 +65,47 @@ extension Reactive where Base: UITextView {
         
         return ControlProperty(values: source, valueSink: bindingObserver)
     }
-    
+
+    /**
+     Reactive wrapper for `delegate` message.
+    */
+    public var didBeginEditing: ControlEvent<()> {
+       return ControlEvent<()>(events: self.delegate.observe(#selector(UITextViewDelegate.textViewDidBeginEditing(_:)))
+            .map { a in
+                return ()
+            })
+    }
+
+    /**
+     Reactive wrapper for `delegate` message.
+     */
+    public var didEndEditing: ControlEvent<()> {
+        return ControlEvent<()>(events: self.delegate.observe(#selector(UITextViewDelegate.textViewDidEndEditing(_:)))
+            .map { a in
+                return ()
+            })
+    }
+
+    /**
+     Reactive wrapper for `delegate` message.
+     */
+    public var didChange: ControlEvent<()> {
+        return ControlEvent<()>(events: self.delegate.observe(#selector(UITextViewDelegate.textViewDidChange(_:)))
+            .map { a in
+                return ()
+            })
+    }
+
+    /**
+     Reactive wrapper for `delegate` message.
+     */
+    public var didChangeSelection: ControlEvent<()> {
+        return ControlEvent<()>(events: self.delegate.observe(#selector(UITextViewDelegate.textViewDidChangeSelection(_:)))
+            .map { a in
+                return ()
+            })
+    }
+
 }
 
 #endif