+++ RxSwift/Observers/AnonymousObserver.swift
@@ -15,7 +15,7 @@ class AnonymousObserver<ElementType> : ObserverBase<ElementType> {
     
     private let _eventHandler : EventHandler
     
-    init(_ eventHandler: EventHandler) {
+    init(_ eventHandler: @escaping EventHandler) {
 #if TRACE_RESOURCES
         let _ = AtomicIncrement(&resourceCount)
 #endif